autenticacion de usuarios con permisos sobre consultas

+++ restauranteServer/routes/dishRouter.js
@@ -18,27 +18,48 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
-    Dishes.create(req.body)
-    .then((dish) => {
-        console.log('Dish Created ', dish);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        Dishes.create(req.body)
+        .then((dish) => {
+            console.log('Dish Created ', dish);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    } 
 })
 .put(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('PUT operation not supported on /dishes');
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        res.statusCode = 403;
+        res.end('PUT operation not supported on /dishes');
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    } 
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.remove({})
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));    
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        Dishes.remove({})
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+    .catch((err) => next(err)); 
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    }    
 });
 
 dishRouter.route('/:dishId')
@@ -53,28 +74,48 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('POST operation not supported on /dishes/'+ req.params.dishId);
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        res.statusCode = 403;
+        res.end('POST operation not supported on /dishes/'+ req.params.dishId);
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    } 
 })
 .put(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findByIdAndUpdate(req.params.dishId, {
-        $set: req.body
-    }, { new: true })
-    .then((dish) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    if(isAdmin){
+        Dishes.findByIdAndUpdate(req.params.dishId, {
+            $set: req.body
+        }, { new: true })
+        .then((dish) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    } 
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findByIdAndRemove(req.params.dishId)
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        Dishes.findByIdAndRemove(req.params.dishId)
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+        } 
 });
 
 dishRouter.route('/:dishId/comments')
@@ -126,26 +167,33 @@ dishRouter.route('/:dishId/comments')
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findById(req.params.dishId)
-    .then((dish) => {
-        if (dish != null) {
-            for (var i = (dish.comments.length -1); i >= 0; i--) {
-                dish.comments.id(dish.comments[i]._id).remove();
+    const isAdmin= authenticate.verifyAdmin(req.user);
+    if(isAdmin){
+        Dishes.findById(req.params.dishId)
+        .then((dish) => {
+            if (dish != null) {
+                for (var i = (dish.comments.length -1); i >= 0; i--) {
+                    dish.comments.id(dish.comments[i]._id).remove();
+                }
+                dish.save()
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);                
+                }, (err) => next(err));
             }
-            dish.save()
-            .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
-            }, (err) => next(err));
-        }
-        else {
-            err = new Error('Dish ' + req.params.dishId + ' not found');
-            err.status = 404;
-            return next(err);
-        }
-    }, (err) => next(err))
-    .catch((err) => next(err));    
+            else {
+                err = new Error('Dish ' + req.params.dishId + ' not found');
+                err.status = 404;
+                return next(err);
+            }
+        }, (err) => next(err))
+        .catch((err) => next(err)); 
+    }else{
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        next(err);
+    }    
 });
 
 dishRouter.route('/:dishId/comments/:commentId')
@@ -180,22 +228,30 @@ dishRouter.route('/:dishId/comments/:commentId')
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
-            if (req.body.rating) {
-                dish.comments.id(req.params.commentId).rating = req.body.rating;
-            }
-            if (req.body.comment) {
-                dish.comments.id(req.params.commentId).comment = req.body.comment;                
-            }
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish._id)
-                .populate('comments.author')
+
+            if(req.user._id.equals(dish.comments.id(req.params.commentId).author._id)){
+                if (req.body.rating) {
+                    dish.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if (req.body.comment) {
+                    dish.comments.id(req.params.commentId).comment = req.body.comment;                
+                }
+                dish.save()
                 .then((dish) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);  
-                })              
-            }, (err) => next(err));
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);  
+                    })              
+                }, (err) => next(err));
+            }else{
+                err = new Error('Can not update or delete the comment posted by other users');
+                err.status = 403;
+                return next(err);
+            }
+
         }
         else if (dish == null) {
             err = new Error('Dish ' + req.params.dishId + ' not found');
@@ -212,20 +268,28 @@ dishRouter.route('/:dishId/comments/:commentId')
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
 
-            dish.comments.id(req.params.commentId).remove();
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish._id)
-                .populate('comments.author')
+            if(req.user._id.equals(dish.comments.id(req.params.commentId).author._id)){
+                dish.comments.id(req.params.commentId).remove();
+                dish.save()
                 .then((dish) => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(dish);  
-                })               
-            }, (err) => next(err));
+                    Dishes.findById(dish._id)
+                    .populate('comments.author')
+                    .then((dish) => {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(dish);  
+                    })               
+                }, (err) => next(err));
+            }else{
+                err = new Error('Can not update or delete the comment posted by other users');
+                err.status = 403;
+                return next(err);
+            }
+            
         }
         else if (dish == null) {
             err = new Error('Dish ' + req.params.dishId + ' not found');
